agent/requirements.txt
fastapi
uvicorn[standard]
httpx
openai
langfuse
python-dotenv
pydantic
pydantic-settings
Pillow>=10.0.0
playwright>=1.40.0
